Move tests as well
